Feature: tag 가중치 데이터 타입 수정 (#255)

+++ tlog-was/src/main/java/com/se/Tlog/domain/Travel/controller/dto/AddFixedTagDto.java
@@ -8,6 +8,6 @@ public record AddFixedTagDto(
         String tagId,
         
         @Schema(description = "태그 가중치")
-        int tagWeight) {
+        double tagWeight) {
 
 }
